requirements.txt
fastapi==0.95.0
sqlalchemy==1.4.0
slowapi==0.1.4
python-dotenv==0.19.2
pydantic[email]==1.9.0
python-jose==3.3.0
httpx==0.23.0
passlib==1.7.4
uvicorn==0.22.0
cryptography==38.0.0
itsdangerous==2.1.2
authlib==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1